Add comparison plots with Avg. Obs

+++ getStatsTva.py
@@ -56,22 +56,35 @@ def mainProg(prjName=None,obsPfx=None,preCur1=None,preCur2=None,prePfx2=None):
   avgObsConn.row_factory = sqlite3.Row
   avgObsCur = avgObsConn.cursor()
   ccolor  = [ 'red', 'blue', 'green', 'cyan', 'magenta', 'lime', 'olive', 'maroon', 'teal', 'purple', 'aqua', 'navy', 'fuchsia']
-      
+  
   for idt,dist in enumerate(distance): 
+    if preCur2 is None:
+      if prePfx2 is None:
+        print "Must provide prefix for SCICHEM-01 prediction"
+        return
+      # Predictions SCICHEM-01
+      pre2DbName =  prePfx2 + '_' + str(dist) + 'km' + '.csv.db'       
+      print 'Predicted DB SCICHEM99', pre2DbName
+      preConn2 = sqlite3.connect(pre2DbName)
+      preConn2.row_factory = sqlite3.Row
+      preCur2 = preConn2.cursor()
+      print preConn2
+    
     lArc = 2.*dist*np.pi/180.
+    
     tblName = "dist" + str(dist)
     
-    avgObsCur.execute('DROP table if exists %s'%tblName)
-    createStr = "CREATE table %s( TRID integer,\
-                TRdist integer, PlumeKM real, species varchar(8), conc real)"%tblName
-    avgObsCur.execute(createStr)
+    if not os.path.exists(avgObsDb):   
+      avgObsCur.execute('DROP table if exists %s'%tblName)
+      createStr = "CREATE table %s( TRID integer,\
+                  TRdist integer, PlumeKM real, species varchar(8), conc real)"%tblName
+      avgObsCur.execute(createStr)
     
     for varName in varNames:
        # Set Omax to zero for max plume no.
       if varName == 'SO2':
         oMax  = [0 for i in range(13)]
-
-      # Observations
+              
       # Observed data
       if "tva_990715" in prjName:
         if dist == 16:
@@ -99,73 +112,43 @@ def mainProg(prjName=None,obsPfx=None,preCur1=None,preCur2=None,prePfx2=None):
         if dist == 89:
           pls = [12] 
         
-        for ipl in pls:
-          
-          obsDbName = obsPfx + str(dist) + 'km_obs' + str(ipl) + '.csv.db'
-          print '\n obsDbName = ',obsDbName
-          obsConn = sqlite3.connect(obsDbName)
-          obsConn.row_factory = sqlite3.Row
-          obsCur = obsConn.cursor()
-          # Observations
-          if "tva_990706" in prjName:
-            obsQry = 'select TRID, CAST(plumeKM as real), ' + varName + '/1000 from dataTable where plumeKM > -9998.'
-          else:
-            obsQry = 'select TRID, CAST(plumeKM as real), ' + varName + ' from dataTable where plumeKM > -9998.'
-          print obsQry
-          obsArray = utilDb.db2Array(obsCur, obsQry)
-          if ipl == min(pls):
-            plt.clf()
-            plt.hold(True)
-          plt.scatter(obsArray[:,1],obsArray[:,2],color=ccolor[ipl],marker='o')
-          if varName == 'SO2':
-             oMax[ipl] = np.where(obsArray[:,2] == obsArray[:,2].max())[0][0]
-             print 'Max Obs SO2 at x index = ', oMax[ipl],' with x, value = ',obsArray[oMax[ipl],1],obsArray[oMax[ipl],2]
-
-          maxDist = obsArray[oMax[ipl],1]
-          for i in range(len(obsArray)):
-            # Set x values where SO2 max to obs x value for plume ip1
-            obsArray[i,1] = obsArray[i,1] - maxDist 
-          plt.scatter(obsArray[:,1],obsArray[:,2],color=ccolor[ipl],marker='d')
-          inStr = "INSERT into %s values("%tblName
-          for myrow in obsArray:
-            insertStr = inStr + '%d, %5.2f, %5.3f, "%s", %g)'%(myrow[0],dist,myrow[1],varName,myrow[2])
-            #print insertStr
-            avgObsCur.execute(insertStr) 
-          #obsArray = obsCur.execute(obsQry)
+        if not os.path.exists(avgObsDb):   
+          for ipl in pls:
             
-          avgObsConn.commit()
-          
-      avgQry = 'select round(plumeKM), avg(conc) from %s where species="%s" and conc >= 0 and plumeKM >-999 group by round(plumeKM)'%(tblName,varName)
-      #print avgQry
-      AveArray2 = utilDb.db2Array(avgObsCur,avgQry)
-      #print AveArray2
-      #plt.axis([0.0,500.0,0,20])
-      plt.plot(AveArray2[:,0], AveArray2[:,1],marker='*')
-      plt.hold(False) 
-      plt.show()
-  sys.exit()
+            obsDbName = obsPfx + str(dist) + 'km_obs' + str(ipl) + '.csv.db'
+            print '\n obsDbName = ',obsDbName
+            obsConn = sqlite3.connect(obsDbName)
+            obsConn.row_factory = sqlite3.Row
+            obsCur = obsConn.cursor()
+            # Observations
+            if "tva_990706" in prjName:
+              obsQry = 'select TRID, CAST(plumeKM as real), ' + varName + '/1000 from dataTable where plumeKM > -9998.'
+            else:
+              obsQry = 'select TRID, CAST(plumeKM as real), ' + varName + ' from dataTable where plumeKM > -9998.'
+            print obsQry
+            obsArray = utilDb.db2Array(obsCur, obsQry)
+            if ipl == min(pls):
+              plt.clf()
+              plt.hold(True)
+            plt.scatter(obsArray[:,1],obsArray[:,2],color=ccolor[ipl],marker='o')
+            if varName == 'SO2':
+               oMax[ipl] = np.where(obsArray[:,2] == obsArray[:,2].max())[0][0]
+               print 'Max Obs SO2 at x index = ', oMax[ipl],' with x, value = ',obsArray[oMax[ipl],1],obsArray[oMax[ipl],2]
+  
+            maxDist = obsArray[oMax[ipl],1]
+            for i in range(len(obsArray)):
+              # Set x values where SO2 max to obs x value for plume ip1
+              obsArray[i,1] = obsArray[i,1] - maxDist 
+            plt.scatter(obsArray[:,1],obsArray[:,2],color=ccolor[ipl],marker='d')
+            inStr = "INSERT into %s values("%tblName
+            for myrow in obsArray:
+              insertStr = inStr + '%d, %5.2f, %5.3f, "%s", %g)'%(myrow[0],dist,myrow[1],varName,myrow[2])
+              #print insertStr
+              avgObsCur.execute(insertStr) 
+            #obsArray = obsCur.execute(obsQry)
+              
+            avgObsConn.commit()
                 
-  for idt,dist in enumerate(distance): 
-
-    if preCur2 is None:
-      if prePfx2 is None:
-        print "Must provide prefix for SCICHEM-01 prediction"
-        return
-      # Predictions SCICHEM-01
-      pre2DbName =  prePfx2 + '_' + str(dist) + 'km' + '.csv.db'       
-      print 'Predicted DB SCICHEM99', pre2DbName
-      preConn2 = sqlite3.connect(pre2DbName)
-      preConn2.row_factory = sqlite3.Row
-      preCur2 = preConn2.cursor()
-      print preConn2
-    
-    lArc = 2.*dist*np.pi/180.
-    
-    
-    for varName in varNames:
-
-      print '\n',dist,' km, ',varName
-      
       # Prediction query
       if varName == "NOx":
         preQry1 = 'select xSmp,Sum(Value) from samTable a,smpTable p where a.colNo=p.colNo'
@@ -197,58 +180,30 @@ def mainProg(prjName=None,obsPfx=None,preCur1=None,preCur2=None,prePfx2=None):
         # Set x index where SO2 is max. Same for all obs plumes
         iMax2 = np.where(preArray2[:,1] == preArray2[:,1].max())[0][0]
                    
-      # Set Omax to zero for max plume no.
-      if varName == 'SO2':
-        oMax  = [0 for i in range(12)]
-
-      figSave = 0
-      #obsAvg = 
-           
-      #insertQry = 'Insert into speciesConc ('+ ipl, obsArray[0],obsArray[1],obsArray[2], obsArray[4]
-      #print insertQry
-      if varName == 'SO2':
-        oMax[ipl] = np.where(obsArray[:,1] == obsArray[:,1].max())[0][0]
-        print 'Max Obs SO2 at x index = ', oMax[ipl],' with x, value = ',obsArray[oMax[ipl],0],obsArray[oMax[ipl],1]
-
-
-      print '\nFor varName %s, plume no. %d,x index = %d, x = %f ,conc = %f\n'%\
-            (varName,ipl,oMax[ipl],obsArray[oMax[ipl],0],obsArray[oMax[ipl],1])
+   
+      avgQry = 'select round(plumeKM), avg(conc) from %s where species="%s" and conc >= 0 and plumeKM >-999 group by round(plumeKM)'%(tblName,varName)
+      #print avgQry
+      obsArray = utilDb.db2Array(avgObsCur,avgQry)
+      #plt.plot(AveArray2[:,0], AveArray2[:,1],marker='*')
+     
+      # Align the prediction1 and observed centerlines
+      for i in range(len(preArray1)):
+        preArray1[i,0] = float(i-iMax1)*lArc      
+      # Set x values where SO2 max to obs x value
+      for i in range(len(preArray2)):
+        preArray2[i,0] = float(i-iMax2)*lArc      
       
-      if figSave:
-        # Align the prediction1 and observed centerlines
-        for i in range(len(preArray1)):
-          # Set x values where SO2 max to obs x value for plume ip1
-          preArray1[i,0] = float(i-iMax1)*lArc + obsArray[oMax[ipl],0]
-          if i == iMax1:
-            print 'Adjusted preArray1 x,c = ',preArray1[i,:] 
-          
-
-        # Align the prediction2 and observed centerlines
-
-        for i in range(len(preArray2)):
-          # Set x values where SO2 max to obs x value for plume ip1
-          preArray2[i,0] = float(i-iMax2)*lArc + obsArray[oMax[ipl],0] 
-          if i == iMax2:
-            print 'Adjusted preArray2 x,c = ',preArray2[i,:]
-         
-        # Plot
-        
-        #figName  = str(dist) +'km_' +varName +'_'+ str(times[idt]) + 'hr_obs' + str(ipl) + '.png'
-        figName  = str(dist) +'km_' +varName +'_'+ str(times[idt]) + 'hr.png'
-        #figTitle = '%s (@ %d km for plume %d)'%(varName,dist,ipl)
-        figTitle = '%s (@ %d km)'%(varName,dist)
-        
-        print figName, varName, str(ipl)
-
-        print varName , figSave, ipl, len(pls), 
-        ##pltCmpConc(dist, varName, obsArray, preArray1, preArray2, figTitle, figName, figSave)
-        
-
-        obsConn.close()
+      # Plot
+      figName  = str(dist) +'km_' +varName +'_'+ str(times[idt]) + 'hr.png'
+      figTitle = '%s (@ %d km)'%(varName,dist)
+      
+      pltCmpConc(dist, varName, obsArray, preArray1, preArray2, figTitle, figName)
+    
     if prePfx2 is not None:
       print 'preConn2 is still open'
       #preConn2.close()
-          
+  avgObsConn.close()
+       
   return
 
 def initDb():
@@ -285,7 +240,7 @@ def insertDb(dbCur,nCol,colTypes,colValues):
   dbCur.execute(insertStr)
   return
           
-def pltCmpConc(dist, varName, obsData, preData1, preData2, figTitle, figName, figSave):
+def pltCmpConc(dist, varName, obsData, preData1, preData2, figTitle, figName):
   #import pdb; pdb.set_trace()
     # Set up plot parameters
   params1 = {'axes.labelsize': 10, 'text.fontsize': 10, 'xtick.labelsize': 10,
@@ -295,8 +250,6 @@ def pltCmpConc(dist, varName, obsData, preData1, preData2, figTitle, figName, fi
   plt.rcParams.update(params1)
   print figName
   fig = plt.figure()
-  if figSave == 1:
-    plt.clf()
   fig.hold(True)
   lgndList = []
   ax = fig.add_subplot(111)
@@ -304,11 +257,11 @@ def pltCmpConc(dist, varName, obsData, preData1, preData2, figTitle, figName, fi
   if varName == 'O3':
     fac = 1
     if obsData is not None:
-     cO = max(obsData[-1,1],obsData[0,1])
+      cO = max(obsData[-1,1],obsData[0,1])
     if preData1 is not None:
       c1 = max(preData1[-1,1],preData1[0,1])
     if preData2 is not None:
-     c2 = max(preData2[-1,1],preData2[0,1])
+      c2 = max(preData2[-1,1],preData2[0,1])
   else:
     fac = 1
     if obsData is not None:
@@ -332,27 +285,21 @@ def pltCmpConc(dist, varName, obsData, preData1, preData2, figTitle, figName, fi
     LhP2 = plt.plot(preData2[:,0],C,linestyle='-',color='cyan',marker='^',markersize=6,markerfacecolor='blue')
     LkP2 = 'SCICHEM-99'
     lgndList.append(LkP2)
-  if figSave:
-    plt.ylabel('Perturbation Concentration (ppm)')
-    plt.xlabel('Cross plume distance (km)')
-    plt.title(figTitle)
+  plt.ylabel('Perturbation Concentration (ppm)')
+  plt.xlabel('Cross plume distance (km)')
+  plt.title(figTitle)
   if dist < 21:
     plt.xlim([-20,20])
   else:
     plt.xlim([-50,50])
   #if varName == 'O3':
   #  plt.ylim([0,0.1])
-  if figSave == 1:
-    plt.legend([LkO,LkP1,LkP2],bbox_to_anchor=(0.02,0.98),loc=2,borderaxespad=0.)
-    lgnd  = plt.gca().get_legend()
-    ltext = lgnd.get_texts()
-    plt.setp(ltext,fontsize=9)
-    fig.hold(False)
-    plt.savefig(figName)
-  else:
-    fig.hold(True)
-  #sys.exit()
-  #plt.show()
+  plt.legend(lgndList,bbox_to_anchor=(0.02,0.98),loc=2,borderaxespad=0.)
+  lgnd  = plt.gca().get_legend()
+  ltext = lgnd.get_texts()
+  plt.setp(ltext,fontsize=9)
+  fig.hold(False)
+  plt.savefig(figName)
   return
 
 def getSmpDb(prjName):
